html_plugin was trying to rename a non-existent file
 - Fix: Check whether the file exists first and don't try to rename it if
   not.

Version:   2015.0113.6

+++ hpssic/plugins/html_plugin.py
@@ -17,6 +17,7 @@ def main(cfg):
     with open(npath, 'w') as out:
         out.write(rpt)
 
-    os.rename(fpath, opath)
+    if os.path.exists(fpath):
+        os.rename(fpath, opath)
     os.rename(npath, fpath)
     CrawlConfig.log("html_plugin finished")
